refactor(dashboard): extract turn-records domain (A1 phase 2 cont.)
- dashboard-turn-records.ts: TurnBlock/TurnRecordEntry/TurnBlockDiff/
  TurnRecordRow/MemoryOs*/KeeperUserModel*/TurnRecordsResponse types +
  decode* helpers + fetchKeeperTurnRecords + TurnTranscript* +
  fetchKeeperTurnTranscript (RFC-0233). normalizeStringList duplicated
  here (turn-records depends on it; canonical lives in keeper-config).
- dashboard.ts: facade re-export only; consumers unchanged
- ripple import cleanup: decodeTelemetryFreshnessMetadata
- typecheck green (tsc --noEmit)

dashboard.ts 2339 -> 1889 lines after turn-records.
